# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  http.createOrderRequest:
    properties:
      amount_cents:
        minimum: 1
        type: integer
      description:
        type: string
      user_id:
        type: string
    required:
    - user_id
    type: object
...
  payments.ConfirmPaymentInput:
    properties:
      amount_cents:
        type: integer
      charge_id:
        type: string
      method:
        type: string
      paid_at:
        type: string
      receipt_url:
        type: string
      tx_hash:
        type: string
    type: object
  payments.EvidenceInput:
    properties:
      file_url:
        type: string
      note:
        type: string
      tx_hash:
        type: string
    type: object
  problem.BaseErr:
    properties:
      details:
        description: Details contém informações adicionais opcionais sobre o erro.
      error:
        description: Title é o título legível do tipo de problema.
        type: string
      errorDescription:
        description: ErrorDescription é uma explicação específica desta ocorrência.
        type: string
      instance:
        description: Instance é o URI da requisição que originou o erro.
        type: string
      statusCode:
        description: StatusCode é o código HTTP associado ao problema.
        type: integer
      timestamp:
        description: Timestamp indica quando o erro ocorreu (UTC, ISO 8601).
        type: string
      type:
        description: Type é um URI que identifica o tipo do problema (único por categoria
          de erro).
        type: string
    type: object
  projects.CreateMilestoneInput:
    properties:
      amount_cents:
        type: integer
      contract_id:
        type: string
      deliverables:
        type: string
      due_date:
        type: string
      title:
        type: string
    type: object
  projects.CreateProjectInput:
    properties:
      contract_id:
        type: string
      name:
        type: string
    type: object
  projects.CreateWorklogInput:
    properties:
      description:
        type: string
      hours:
        type: number
      milestone_id:
        type: string
      worked_at:
        type: string
    type: object
host: localhost:8000
info:
  contact: {}
  description: API modular de pagamentos com autenticação Bearer Token.
  title: AggiPay API
  version: 1.0.0
paths:
...
  /orders:
    post:
      consumes:
      - application/json
      parameters:
      - description: Dados do pedido
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/http.createOrderRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties: true
            type: object
      summary: Cria um pedido de pagamento
      tags:
      - Orders
  /orders/{id}:
    get:
      parameters:
      - description: ID do pedido
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Consulta um pedido pelo ID
      tags:
      - Orders
  /payments:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/problem.BaseErr'
      security:
      - BearerAuth: []
      summary: Listar pagamentos
      tags:
      - Payments
  /payments/{id}/evidence:
    post:
      consumes:
      - application/json
      description: Anexa evidência/comprovante e hash de transação cripto
      parameters:
      - description: ID do pagamento
        in: path
        name: id
        required: true
        type: string
      - description: Dados do comprovante
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/payments.EvidenceInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/problem.BaseErr'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/problem.BaseErr'
      security:
      - BearerAuth: []
      summary: Adicionar comprovante
      tags:
      - Payments
  /payments/confirm:
    post:
      consumes:
      - application/json
      description: Confirma pagamento manual ou por conciliação externa
      parameters:
      - description: Dados do pagamento
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/payments.ConfirmPaymentInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/problem.BaseErr'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/problem.BaseErr'
      security:
      - BearerAuth: []
      summary: Confirmar pagamento
      tags:
      - Payments
  /projects:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/problem.BaseErr'
      security:
      - BearerAuth: []
      summary: Listar projetos
      tags:
      - Projects
    post:
      consumes:
      - application/json
      parameters:
      - description: Dados do projeto
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/projects.CreateProjectInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/problem.BaseErr'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/problem.BaseErr'
      security:
      - BearerAuth: []
      summary: Criar projeto
      tags:
      - Projects
  /projects/{id}/milestones:
    post:
      consumes:
      - application/json
      parameters:
      - description: ID do projeto
        in: path
        name: id
        required: true
        type: string
      - description: Dados do milestone
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/projects.CreateMilestoneInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/problem.BaseErr'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/problem.BaseErr'
      security:
      - BearerAuth: []
      summary: Criar milestone
      tags:
      - Projects
  /projects/{id}/worklogs:
    post:
      consumes:
      - application/json
      parameters:
      - description: ID do projeto
        in: path
        name: id
        required: true
        type: string
      - description: Dados do worklog
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/projects.CreateWorklogInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/problem.BaseErr'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/problem.BaseErr'
      security:
      - BearerAuth: []
      summary: Criar worklog
      tags:
      - Projects
  /webhooks/asaas:
    post:
      consumes:
      - application/json
      description: Recebe eventos de cobrança/pagamento do Asaas
      parameters:
      - description: Evento Asaas
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Webhook Asaas
      tags:
      - Webhooks
  /webhooks/clicksign:
    post:
      consumes:
      - application/json
      description: Recebe eventos de assinatura digital do Clicksign
      parameters:
      - description: Evento Clicksign
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Webhook Clicksign
      tags:
      - Webhooks
securityDefinitions:
  BearerAuth:
    description: 'Informe: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"